+++ settings.gradle
@@ -1,2 +1,2 @@
-rootProject.name = 'CustomMavenBadges'
+rootProject.name = 'MavenHelper'
