+++ src/main/java/me/kingtux/mvnhelper/BadgeUtils.java
@@ -1,4 +1,4 @@
-package me.kingtux.cmb;
+package me.kingtux.mvnhelper;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
