+++ src/main/java/me/kingtux/mvnhelper/Config.java
@@ -1,4 +1,4 @@
-package me.kingtux.cmb;
+package me.kingtux.mvnhelper;
 
 import com.google.gson.JsonObject;
